refactor: 관리자 페이지 CouponApis 폴더 위치 promotion으로 통합합
